fix: navigation과 component 책임 분리
