+++ src/components/common/TabBar.tsx
@@ -1,27 +1,35 @@
 import React from 'react';
 import {BottomTabBarProps} from '@react-navigation/bottom-tabs';
-import {View, TouchableOpacity, Text} from 'react-native';
+import {View, TouchableOpacity, Text, Image} from 'react-native';
+import {HomeStackParamList} from '../../types/StackNavigationType';
+
+type TabBarComponentType = {
+  [route in keyof HomeStackParamList]: {
+    name: string;
+    icon: string;
+  };
+};
+
+const tabBarData: TabBarComponentType = {
+  Feed: {
+    name: '홈',
+    icon: 'https://legacy.reactjs.org/logo-og.png',
+  },
+  MyPage: {
+    name: '마이 페이지',
+    icon: 'https://legacy.reactjs.org/logo-og.png',
+  },
+};
+
+const TabBarIcon = ({icon}: {icon: string}) => {
+  return <Image source={{uri: icon}} style={{width: 24, height: 24}} />;
+};
 
 const TabBar = ({state, descriptors, navigation}: BottomTabBarProps) => {
   return (
     <View style={{flexDirection: 'row'}}>
       {state.routes.map((route, index) => {
         const {options} = descriptors[route.key];
-
-        const Icon = options.tabBarIcon
-          ? options.tabBarIcon({
-              focused: state.index === index,
-              color: '#673ab7',
-              size: 24,
-            })
-          : null;
-
-        const label =
-          options.tabBarLabel !== undefined
-            ? options.tabBarLabel
-            : options.title !== undefined
-            ? options.title
-            : route.name;
 
         const isFocused = state.index === index;
 
@@ -54,9 +62,9 @@ const TabBar = ({state, descriptors, navigation}: BottomTabBarProps) => {
             onPress={onPress}
             onLongPress={onLongPress}
             style={{flex: 1}}>
-            {Icon}
+            <TabBarIcon icon={tabBarData[route.name].icon} />
             <Text style={{color: isFocused ? '#673ab7' : '#222'}}>
-              {typeof label === 'string' ? label : null}
+              {tabBarData[route.name].name}
             </Text>
           </TouchableOpacity>
         );
