+++ src/types/StackNavigationType.ts
@@ -1,24 +1,26 @@
+import {ParamListBase} from '@react-navigation/native';
+
 type StackParamType<T> = {
   screen?: keyof T;
   params?: T[keyof T];
 };
 
-export type HomeStackParamList = {
+export interface HomeStackParamList extends ParamListBase {
   Feed: undefined;
   MyPage: undefined;
-};
+}
 
-export type RegisterStackParamList = {
+export interface RegisterStackParamList extends ParamListBase {
   Signup: undefined;
   Login: undefined;
-};
+}
 
-export type DetailStackParamList = {
+export interface DetailStackParamList extends ParamListBase {
   Store: {storeId: number};
-};
+}
 
-export type RootStackParamList = {
+export interface RootStackParamList extends ParamListBase {
   Home: StackParamType<HomeStackParamList>;
   Register: StackParamType<RegisterStackParamList>;
   Detail: StackParamType<DetailStackParamList>;
-};
+}
